requirements.txt
aenum==3.1.16
arrow==1.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
flatdict==4.0.1
idna==3.10
pycparser==2.22
pydash==8.0.5
PyJWT==2.10.1
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
urllib3==2.5.0
zscaler-sdk-python==1.5.8
